Add endpoint and UI for editing assay version name
Introduces a new API endpoint `save_version_name` in `assays.py` and registers it in `urls.py` to allow updating the version name of an assay version. Updates the configuration template to provide a UI for editing and saving the version name instead of the assay name, and adjusts related JavaScript and form fields accordingly. Also removes debug print statements and cleans up version name display in templates.